pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr"
version = "0.1.0"
description = "TODO."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Marko Rosenmüller" }]
keywords = []
dependencies = [
  "flask==3.1.2",
  "uvicorn==0.40.0",
  "asgiref==3.11.1",
  "openai>=1.0.0",
  "pillow>=11.0.0",
  "pdf2image>=1.17.0",
]

[project.scripts]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["src*"]
exclude = ["tests*", "docs*"]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
markers = [
  "integration: requires a running LLM server at localhost:8080",
]
